yaffs2: disable procfs support
The proc interfaces in the 3.10 kernel have changed in a manner that is not
compatible with yaffs2. As discussed in the following thread:

   http://www.aleph1.co.uk/lurker/message/20130724.003329.c28842f5.en.html

The solution is to temporarily remove procfs support, which allows the FS to
continue to be used.

Based on this patch:

  https://dev.openwrt.org/export/37617/trunk/target/linux/generic/patches-3.10/515-yaffs-3.10-disable-proc-entry.patch

We disable procfs for any kernel greater than 3.9

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>