MXSCM-226 dts: imx: add pm-ignore-notify option for sd3 on mx6sxscm evb
There is no card-detection pad connected for sd3 on the i.mx6 SX
EVB board. The card is assumed to be non-removable, hence, there is
no need to redetect the card during the pm callbacks. This can be
reached by including the pm-ignore-notify option on the usdhc3 device
on the dtb file.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>
[zou:Original patch taken from
git://git.freescale.com/imx/linux-imx.git imx_4.9.11_1.0.0_ga]
Signed-off-by: zou cao <cao.zou@windriver.com>