MLK-14675: ARM: dts: set pll4 rate for audio
There is noise when use ssi master mode. the reason is the ssi rate
is not accurate for ssi master mode, show below.

    pll3_pfd2_508m               0            0   508235294          0 0
       ssi3_sel                  0            0   508235294          0 0
          ssi3_pred              0            0   127058824          0 0
             ssi3_podf           0            0    63529412          0 0
                ssi3             0            0    63529412          0 0
       ssi2_sel                  0            0   508235294          0 0
          ssi2_pred              0            0   127058824          0 0
             ssi2_podf           0            0    63529412          0 0
                ssi2             0            0    63529412          0 0

so we need to switch ssi's parent to pll4 (which is dedicate audio pll),
and set a proper rate for pll4, we select 786432000Hz, which can
generate 32kHz,48kHz,96KHz.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
[zou:Original patch taken from
git://git.freescale.com/imx/linux-imx.git imx_4.9.11_1.0.0_ga]
Signed-off-by: zou cao <cao.zou@windriver.com>