MLK-14700 tty: serial: imx: only enable RTSDEN if have_rtscts
This is very similar to an older patch:

commit e1cbbcd1d1ac ("MLK-12161 tty: serial: imx: only enable RTSD
interrupt for hw flow control")

The difference is that it adds the same check on the suspend path. This
fixes suspend not working on imx6sl-evk when uart wakeup is enabled.
What happens instead is that an interrupt is imediately received on the
suspend path.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
[zou:Original patch taken from
git://git.freescale.com/imx/linux-imx.git imx_4.9.11_1.0.0_ga]
Signed-off-by: zou cao <cao.zou@windriver.com>