iwlwifi: mvm: reset the HW before dumping if HW error is detected
commit 6b54ebf73b3e5adc2cab3abc6cb37ca66fa74ae2 upstream.

If the hardware is stuck, we can't read any of the memory we need to
dump it, so we end up printing only 0xa5a5a5a5, which is useless.

To solve this, poke the hardware by triggering a reset and re-enabling
the clocks if we detect a HW error.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>