FogBugz #492768: i2c: dw: temporary workaround for i2c clocks
commit  84316f46573780c07a4ecb8933b4516d3265a87c from
https://github.com/altera-opensource/linux-socfpga.git

This is a temporary workaround that:
* will not be submitted upstream
* will be reverted when the clock driver is fixed

Currently there are issues with writing to the stratix 10 clock
controller registers.  This commit allows the i2c controller to work
without doing any clock controller register access since we know the
clocks are enabled and we know what frequency they are.

This workaround is enabled by a setting in the stratix 10 dts, so it
will not affect other platforms that don't have this setting.

Signed-off-by: Alan Tull <atull@kernel.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>