ALSA: hda - Add AZX_DRIVER_SKL for simplification
commit a4b4793f640b72af3e2bb2a1ad79725c103d5e40 upstream.

We checked the quirks specific to the recent Intel chips by checking
the PCI IDs manually, but it's becoming messy with lots of IS_SKL()
and other macros, as the amount accumulated.

For simplification, here the new AZX_DRIVER_SKL type is introduced,
and check chip->driver_type instead of the manual PCI ID.  The short
name for this is still "HDA Intel PCH", so that it doesn't break the
existing user-space unnecessarily.

Suggested-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Liwei Song <liwei.song.lsong@gmail.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>