ALSA: hda - Bind with i915 component before codec binding
commit fcc88d91cd36d1343a0ccc09444b21f6b0dad2d8 upstream.

We used a on-demand i915 component binding for IvyBridge and
SandyBridge HDMI codecs, but it has a potential problem of the nested
module loading.  For avoiding that situation, assure the i915 binding
happening at the controller driver level for PCH controller devices,
where the initialization is performed in a detached work, instead of
calling from the codec driver probe.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Liwei Song <liwei.song.lsong@gmail.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>