soc: imx: gpc: adjust GPC_CLK_MAX from 6 to 10
In function "imx_pgc_get_clocks", GPC_CLK_MAX is used to limit
the numbers of clocks which attached to device node "gpc". In
imx6qdl.dtsi, there are 7 clocks which are listed as below for gpc:

               clocks = <&clks IMX6QDL_CLK_GPU3D_CORE>,
                     <&clks IMX6QDL_CLK_GPU3D_SHADER>,
                     <&clks IMX6QDL_CLK_GPU2D_CORE>,
                     <&clks IMX6QDL_CLK_GPU2D_AXI>,
                     <&clks IMX6QDL_CLK_OPENVG_AXI>,
                     <&clks IMX6QDL_CLK_VPU_AXI>,
                     <&clks IMX6QDL_CLK_IPG>;

With this it will throw out an error as below:
imx-gpc 20dc000.gpc: more than 6 clocks

So we need to adjust it from 6 to 10.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>